pyproject.toml
[tool.poetry]
name = "monetify"
version = "0.1.0"
description = ""
authors = ["Erkinumur <erkin.umur94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
drf-firebase-auth = "^1.0.0"
loguru = "^0.5.3"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
